+++ ors-engine/src/main/java/org/heigit/ors/routing/graphhopper/extensions/flagencoders/FootFlagEncoder.java
@@ -116,7 +116,8 @@ public abstract class FootFlagEncoder extends com.graphhopper.routing.util.FootF
                 "living_street",
                 "track",
                 "residential",
-                "service"
+                "service",
+                "corridor"
         ));
 
         avoidHighwayTags.addAll(Arrays.asList(